CSS escapes: Make it possible to escape special characters
This patch adds a checkbox that allows the user to escape any non-printable ASCII characters as well as non-ASCII characters, both in the CSS selector as in the JavaScript output.

In case the checkbox is checked, and supplementary (non-BMP) Unicode characters are used, warn the user that they should be used unescaped because of a WebKit bug.

E.g. http://mothereff.in/css-escapes#01a%5C%2F%222b3c, http://mothereff.in/css-escapes#11a%F0%9D%8C%86wut